+++ Eis/src/Eis/Debug/Instrumentor.h
@@ -8,6 +8,8 @@
 #include <algorithm>
 #include <thread>
 
+// TODO: populate engine with profiling
+
 namespace Eis
 {
 	struct ProfileResult
@@ -110,7 +112,7 @@ namespace Eis
 			int64_t start = std::chrono::time_point_cast<std::chrono::microseconds>(m_StartTimepoint).time_since_epoch().count();
 			int64_t end = std::chrono::time_point_cast<std::chrono::microseconds>(endTimepoint).time_since_epoch().count();
 
-			uint32_t threadId = std::hash<std::thread::id>{}(std::this_thread::get_id());
+			uint32_t threadId = (uint32_t)std::hash<std::thread::id>{}(std::this_thread::get_id());
 			Instrumentor::Get().WriteProfile({m_Name, start, end, threadId});
 
 			m_Stopped = true;
